backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
contourpy==1.3.1
cycler==0.12.1
Django==5.1.5
django-extensions==3.2.3
djangorestframework==3.15.2
fonttools==4.56.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.2
ollama==0.4.7
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psycopg2==2.9.10
pydantic==2.11.1
pydantic_core==2.33.0
pydot==3.0.4
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
threadpoolctl==3.5.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.1
